refactor: deduplicate bfloat16 ULP bit-manipulation helpers in eltwise tests
Move five identical helper functions (float_to_bf16_bits, bf16_bits_to_float,
is_bf16_denormal, bf16_daz_normalize, bf16_value_order_index_daz) that were
copy-pasted verbatim into four eltwise ULP test files into the existing
eltwise_test_utils.py, and import them instead.

Pure relocation — function bodies are byte-identical to the originals.
