+++ tests/ttnn/unit_tests/operations/eltwise/test_tanh_bw_ulp.py
@@ -23,62 +23,17 @@ Per tech_reports/Handling_Special_Value/special_values.md: "denormals | all | 0x
 Run: pytest tests/ttnn/unit_tests/operations/eltwise/test_tanh_bw_ulp.py -v -s
 """
 
-import struct
 import pytest
 import torch
 import ttnn
 from loguru import logger
 from mpmath import mp, cosh as mp_cosh
-
-
-def float_to_bf16_bits(f: float) -> int:
-    """Convert float to BFloat16 bit representation."""
-    f32_bits = struct.unpack(">I", struct.pack(">f", f))[0]
-    return f32_bits >> 16
-
-
-def bf16_bits_to_float(bits: int) -> float:
-    """Convert BFloat16 bits to float."""
-    f32_bits = bits << 16
-    return struct.unpack(">f", struct.pack(">I", f32_bits))[0]
-
-
-def is_bf16_denormal(bits: int) -> bool:
-    """Check if BF16 bits represent a denormal (subnormal) value."""
-    exp = (bits >> 7) & 0xFF
-    mantissa = bits & 0x7F
-    return (exp == 0) and (mantissa != 0)
-
-
-def bf16_daz_normalize(bits: int) -> int:
-    """Apply DAZ (Denormals-Are-Zero) normalization to BF16 bits."""
-    if is_bf16_denormal(bits):
-        return 0x0000
-    if bits == 0x8000:  # -0 -> +0
-        return 0x0000
-    return bits
-
-
-def bf16_value_order_index_daz(bits: int) -> int:
-    """Calculate the value order index for a BFloat16 value with DAZ."""
-    bits = bf16_daz_normalize(bits)
-
-    exp = (bits >> 7) & 0xFF
-    mantissa = bits & 0x7F
-    if exp == 0xFF and mantissa != 0:
-        return -1  # NaN
-    if bits == 0x7F80:
-        return 65281  # +inf
-    if bits == 0xFF80:
-        return -1  # -inf
-    if bits == 0x0000:
-        return 32640  # Zero
-
-    if bits & 0x8000:
-        magnitude = bits & 0x7FFF
-        return 0x7F7F - magnitude
-    else:
-        return 32640 + bits - 0x007F
+from tests.ttnn.unit_tests.operations.eltwise.eltwise_test_utils import (
+    float_to_bf16_bits,
+    bf16_bits_to_float,
+    bf16_daz_normalize,
+    bf16_value_order_index_daz,
+)
 
 
 def ulp_distance_bf16_daz(a: float, b: float) -> int:
